add int casting on ceil/floor in for loop

+++ assessment/interpret5.php
@@ -190,9 +190,9 @@ function interpretline($str,$anstype,$countcnt,$included_qs=[]) {
 					$bits = array( "if (is_nan({$forcond[2]}) || is_nan({$forcond[1]})) {echo 'part of for loop is not a number';} else {
 						\$_forends[{$countcnt}] = [round(floatval({$forcond[1]}),4), round(floatval({$forcond[2]}),4)];
 						if (\$_forends[{$countcnt}][0] <= \$_forends[{$countcnt}][1]) {
-							for ({$forcond[0]}=ceil(\$_forends[{$countcnt}][0]),\$forloopcnt[{$countcnt}]=0;{$forcond[0]}<=floor(\$_forends[{$countcnt}][1]) && \$forloopcnt[{$countcnt}]<1000; {$forcond[0]}++, \$forloopcnt[{$countcnt}]++) {".$todo."};
+							for ({$forcond[0]}=(int)ceil(\$_forends[{$countcnt}][0]),\$forloopcnt[{$countcnt}]=0;{$forcond[0]}<=(int)floor(\$_forends[{$countcnt}][1]) && \$forloopcnt[{$countcnt}]<1000; {$forcond[0]}++, \$forloopcnt[{$countcnt}]++) {".$todo."};
 						} else {
-							for ({$forcond[0]}=floor(\$_forends[{$countcnt}][0]),\$forloopcnt[{$countcnt}]=0;{$forcond[0]}>=ceil(\$_forends[{$countcnt}][1]) && \$forloopcnt[{$countcnt}]<1000; {$forcond[0]}--, \$forloopcnt[{$countcnt}]++) {".$todo."};
+							for ({$forcond[0]}=(int)floor(\$_forends[{$countcnt}][0]),\$forloopcnt[{$countcnt}]=0;{$forcond[0]}>=(int)ceil(\$_forends[{$countcnt}][1]) && \$forloopcnt[{$countcnt}]<1000; {$forcond[0]}--, \$forloopcnt[{$countcnt}]++) {".$todo."};
 						} 
 						if (\$forloopcnt[{$countcnt}]>=1000) {echo \"for loop exceeded 1000 iterations - giving up\";}}");
 				} else {
